+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Yii2 multiple input change log
 2.22.0 (in development)
 =======================
 - Ignore dev files in zip distribution (sup-ham)
+- #292 Fixed tests for last PHPUnit
+- Added support prepare values of attributes with same name as the relation
 
 2.21.4
 ======